refactor: standardize React component naming convention
